feat: split automation workspaces
