+++ langwatch/src/pages/[project]/automations/activity.tsx
@@ -1,0 +1,1 @@
+export { default } from "../automations";
