+++ langwatch/src/pages/[project]/automations/alerts.tsx
@@ -1,0 +1,1 @@
+export { default } from "../automations";
