+++ langwatch/src/pages/[project]/automations/schedules.tsx
@@ -1,0 +1,1 @@
+export { default } from "../automations";
